refactoring: rudiments are removed
