+++ Assets/OneLine/Editor/Drawers/Directory/DirectoryDrawer.cs
@@ -15,22 +15,5 @@ namespace OneLine {
             return property.GetChildren();
         }
 
-        protected override void DrawField(Rect rect, SerializedProperty property) {
-            DrawHighlight(rect, property);
-            getDrawer(property).Draw(rect, property);
-            DrawTooltip(rect, property);
-        }
-
-        protected void DrawTooltip(Rect rect, SerializedProperty child) {
-            string tooltip = child.displayName;
-
-            var attribute = child.GetCustomAttribute<TooltipAttribute>();
-            if (attribute != null) {
-                tooltip = attribute.tooltip;
-            }
-
-            EditorGUI.LabelField(rect, new GUIContent("", tooltip));
-        }
-
     }
 }
